Flytt Datoformat-enum til fil med andre dato-funksjoner
