+++ src/frontend/komponenter/Fagsak/Behandlingsresultat/Valutakurs/ValutakursTabellRad.tsx
@@ -10,7 +10,8 @@ import {
 } from '../../../../context/Valutakurs/ValutakursSkjemaContext';
 import type { IBehandling } from '../../../../typer/behandling';
 import type { IRestValutakurs } from '../../../../typer/eøsPerioder';
-import { Datoformat, formaterIsoDato, lagPersonLabel } from '../../../../utils/formatter';
+import { Datoformat } from '../../../../utils/dato';
+import { formaterIsoDato, lagPersonLabel } from '../../../../utils/formatter';
 import { StatusBarnCelleOgPeriodeCelle } from '../EøsPeriode/fellesKomponenter';
 
 interface IProps {
