+++ src/frontend/komponenter/Fagsak/Høyremeny/Høyremeny.tsx
@@ -9,7 +9,8 @@ import { hentDataFraRessursMedFallback, RessursStatus } from '@navikt/familie-ty
 
 import { useBehandling } from '../../../context/behandlingContext/BehandlingContext';
 import type { ILogg } from '../../../typer/logg';
-import { formaterIsoDato, Datoformat } from '../../../utils/formatter';
+import { Datoformat } from '../../../utils/dato';
+import { formaterIsoDato } from '../../../utils/formatter';
 import Hendelsesoversikt from '../../Felleskomponenter/Hendelsesoversikt/Hendelsesoversikt';
 import type { Hendelse } from '../../Felleskomponenter/Hendelsesoversikt/typer';
 import Behandlingskort from '../Behandlingskort/Behandlingskort';
@@ -22,7 +23,7 @@ const ToggleVisningHøyremeny = styled(Button)`
     width: 34px;
     min-width: 34px;
     height: 34px;
-    padding: 0px;
+    padding: 0;
     border-radius: 50%;
     filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
     background-color: ${ASurfaceDefault};
