+++ src/frontend/komponenter/Fagsak/Saksoversikt/Saksoversikt.tsx
@@ -22,10 +22,9 @@ import {
 import type { IMinimalFagsak } from '../../../typer/fagsak';
 import { FagsakStatus } from '../../../typer/fagsak';
 import { Vedtaksperiodetype } from '../../../typer/vedtaksperiode';
-import { isoStringTilDate } from '../../../utils/dato';
+import { Datoformat, isoStringTilDate } from '../../../utils/dato';
 import { periodeOverlapperMedValgtDato } from '../../../utils/dato';
 import { hentAktivBehandlingPåMinimalFagsak } from '../../../utils/fagsak';
-import { Datoformat } from '../../../utils/formatter';
 import { Infotrygdtabeller } from '../../Infotrygd/Infotrygdtabeller';
 import { useInfotrygdRequest } from '../../Infotrygd/useInfotrygd';
