+++ src/frontend/komponenter/Fagsak/Vilkårsvurdering/GeneriskVilkår/VilkårTabellRad.tsx
@@ -15,8 +15,8 @@ import VilkårResultatIkon from '../../../../ikoner/VilkårResultatIkon';
 import type { IGrunnlagPerson } from '../../../../typer/person';
 import type { IVilkårConfig, IVilkårResultat } from '../../../../typer/vilkår';
 import { Resultat, resultatVisningsnavn } from '../../../../typer/vilkår';
-import { isoDatoPeriodeTilFormatertString } from '../../../../utils/dato';
-import { Datoformat, formaterIsoDato } from '../../../../utils/formatter';
+import { Datoformat, isoDatoPeriodeTilFormatertString } from '../../../../utils/dato';
+import { formaterIsoDato } from '../../../../utils/formatter';
 import { alleRegelverk } from '../../../../utils/vilkår';
 
 interface IProps {
