+++ src/frontend/komponenter/Fagsak/Vilkårsvurdering/Vilkårsvurdering.tsx
@@ -23,7 +23,8 @@ import type { IBehandling } from '../../../typer/behandling';
 import { BehandlingSteg, BehandlingÅrsak } from '../../../typer/behandling';
 import type { IAnnenVurdering, IVilkårResultat } from '../../../typer/vilkår';
 import { annenVurderingConfig, vilkårConfig } from '../../../typer/vilkår';
-import { Datoformat, formaterIsoDato } from '../../../utils/formatter';
+import { Datoformat } from '../../../utils/dato';
+import { formaterIsoDato } from '../../../utils/formatter';
 import { erProd } from '../../../utils/miljø';
 import { hentFrontendFeilmelding } from '../../../utils/ressursUtils';
 import Skjemasteg from '../../Felleskomponenter/Skjemasteg/Skjemasteg';
