+++ src/frontend/komponenter/Felleskomponenter/Datovelger/Månedvelger.tsx
@@ -7,8 +7,7 @@ import { MonthPicker, useMonthpicker } from '@navikt/ds-react';
 import type { Felt } from '@navikt/familie-skjema';
 
 import { senesteRelevanteDato, tidligsteRelevanteDato } from './utils';
-import { dagensDato, dateTilFormatertString } from '../../../utils/dato';
-import { Datoformat } from '../../../utils/formatter';
+import { dagensDato, dateTilFormatertString, Datoformat } from '../../../utils/dato';
 
 interface IProps {
     felt: Felt<Date | undefined>;
