+++ src/frontend/utils/dato/dato.ts
@@ -3,8 +3,6 @@ import { format, isValid, parseISO, startOfToday } from 'date-fns';
 import type { FeltState } from '@navikt/familie-skjema';
 import { feil, ok } from '@navikt/familie-skjema';
 
-import { Datoformat } from '../formatter';
-
 export type IsoDatoString = string; // Format YYYY-MM-DD (ISO)
 export type IsoMånedString = string; // Format YYYY-MM (ISO)
 
@@ -13,6 +11,21 @@ export const dagensDato = startOfToday();
 export const tidenesMorgen = new Date(1000, 1, 1);
 
 export const tidenesEnde = new Date(3000, 1, 1);
+
+export enum Datoformat {
+    DATO = 'dd.MM.yyyy',
+    DATO_FORKORTTET = 'dd.MM.yy',
+    DATO_FORLENGET = 'PPP',
+    DATO_FORLENGET_MED_TID = 'PPPp',
+    ISO_MÅNED = 'yyyy-MM',
+    ISO_DAG = 'yyyy-MM-dd',
+    DATO_TID = 'dd.MM.yy HH:mm',
+    DATO_TID_SEKUNDER = 'dd.MM.yy HH:mm:ss',
+    MÅNED_ÅR = 'MM.yyyy',
+    MÅNED_ÅR_NAVN = 'MMMM yyyy',
+    MÅNED_ÅR_KORTNAVN = 'MMM yyyy',
+    MÅNED_NAVN = 'MMM',
+}
 
 interface DateTilFormatertStringProps {
     date?: Date;
